ref(callout): unsubscribe only just when there is an existing subscription
The callout can be configured to not be able to close on backdrop click. In this case the directive does not get the backdropClicked Observable from the OverlayRef and does not subscribe to it, so without this existence check there would be a runtime TypeError when the CalloutConfig’s canDismiss property is set to false.